Update privacy-policy.md
Made a small change. Changed from "We will not always store cookies in your browser, but when we will do it we'll try the best." to "We will not always store cookies in your browser, but when we do we will do our best to protect your personal data." 

I felt like this made more sense for what you were trying to say!